# Prevent excessive trajectory slicing