@sroet 2021-02-19
@dwhswenson this is ready for a review. As you probably know this code better than I do, are there any other places where a similar strategy might be beneficial? (looking at it quickly, I can probably mimic this pretty easy for `_find_subtraj_first`, but that function has not popped up yet when I try to kill on Trajectory generation)