@sroet 2021-02-21
> What you have here works, but I think there's a much easier way (as opposed to the exttraj iterator, etc.) Wouldn't this do the job:
```python
# subtraj = traj[slice(start, stop)]
# replace with
subtraj = list(traj)[slice(start, stop)]
```
> I think all we need is an iterable of snapshots, so this would create that with very little change to the code.

I did do that in https://github.com/openpathsampling/openpathsampling/pull/978/commits/1c156433c6f7c2a1d5084b770aef970d913ea258 , this would possibly break other custom ensembles, but the workaround is generally easy and the limited change/robustness in this code is worth it in my opinion. 

As ensembles are not fully covered under coverage, please do have a **really** good look at this PR. Here is a new set of numbers from this PR:

|  | master| 1c15643 | b89bce9 |
| --- | --- | --- | --- |
||71186|5850| 62 |
||75610|3257| 114 |
||76504|7526| 86 |
||67516|6842| 99 |
||71031|4789| 47 |
||67648|9378| 33 |
|*mean*| 71186 | 6274 | 74

Currently this PR prevents **99.9%** of all created trajectories in 10 mc steps And these number seem reasonable for 44 ensembles and 10 mc steps