@sroet 2021-02-23
@dwhswenson implemented the `proxy_loading` as you advised, please have another look.

I updated the table in https://github.com/openpathsampling/openpathsampling/pull/978#issuecomment-782896532 with the new data. (still sub `10` Trajectories per md step)