@sroet 2021-02-23
> It might be worth just having a function called _get_list_traj which does this getattr/list pair of operations. I suspect CodeClimate will complain about repeated code here, and since the reason for the need is obscure enough, it might be worth it to have a single location with a clearer docstring explaining what's going on. That's optional though.

Yeah, looking back at it that is the cleaner and more maintainable solution, will refactor that now.