@sroet 2021-02-23
@dwhswenson refactored the code, it still results in numbers in the same order as 1cff417 . Please have a look if you think the docstring is descriptive enough.

By the way, thank you for the (multiple) in-depth reviews.