@dwhswenson 2021-02-24
> Please have a look if you think the docstring is descriptive enough.

LGTM. If anything, it's probably more detailed than needed!

> By the way, thank you for the (multiple) in-depth reviews.

No problem, and thank *you* for making the improvements!

Merging this one now.